move api related class to api package
