+++ controller/src/main/java/ai/starwhale/mlops/api/protocol/ResponseMessage.java
@@ -5,9 +5,8 @@
  * in accordance with the terms of the license agreement you entered into with StarWhale.com.
  */
 
-package ai.starwhale.mlops.common;
+package ai.starwhale.mlops.api.protocol;
 
-import ai.starwhale.mlops.constants.Code;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import lombok.Data;
